Refactor(#102): 서버에서 Prefetching 후 데이터 de/hydration
